refactor: domain 패키지 분리(test 동일)
domain.car
 - CarName
 - CarPosition
domain.race
 - RaceCount
 - RaceNumber
