+++ src/main/java/racinggame/domain/car/CarPosition.java
@@ -1,4 +1,4 @@
-package racinggame.domain;
+package racinggame.domain.car;
 
 public class CarPosition {
     private int position;
