+++ src/main/java/racinggame/domain/race/RaceCount.java
@@ -1,4 +1,4 @@
-package racinggame.domain;
+package racinggame.domain.race;
 
 import racinggame.utils.ValidationUtils;
