+++ src/test/java/racinggame/domain/car/CarNameTest.java
@@ -1,4 +1,4 @@
-package racinggame.domain;
+package racinggame.domain.car;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
