+++ src/test/java/racinggame/domain/car/CarPositionTest.java
@@ -1,7 +1,6 @@
-package racinggame.domain;
+package racinggame.domain.car;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
