+++ src/test/java/racinggame/domain/race/RaceCountTest.java
@@ -1,4 +1,4 @@
-package racinggame.domain;
+package racinggame.domain.race;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
